LocationPage: Simplify StopMap component
StopMap component can load as soon as the page loads, but currently, it
is only loaded after the latitude/longitude is fetched. It is a big
change to do that. As a first stop, this commit removes the
latitude/longitude dependency.

In this commit, StopMap component is loaded with a view of Manhattan
with zoom level 11 as default, and when the GeolocateControl discovers
the current location, it will automatically zoom-in to the current
location at zoom 14.

This commit:
- Removes latitude/longitude props from StopMap component
- Adds MAP_INITIAL_VIEW_STATE constant for Manhattan center
  (40.782864, -73.965355) at zoom 11
- Renames OpenMapTilesStyle to OPEN_FREE_MAP_TILES_STYLE for consistency
- Updates LocationPage to pass only stops prop to StopMap